pyproject.toml
[project]
name = "data-science-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk>=1.14",
    "immutabledict>=4.2.1",
    "sqlglot>=26.10.1",
    "db-dtypes>=1.4.2",
    "regex>=2024.11.6",
    "tabulate>=0.9.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "absl-py>=2.2.2",
    "pydantic>=2.11.3",
    "pandas>=2.3.0",
    "numpy>=2.3.1",
    "toolbox-core>=0.3.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "pg8000>=1.31.2",
    "litellm>=1.80.11",
]
